v1.0.2, added subservices main image url.

+++ src/main/java/com/marcebanizi/backend/model/ModeloServicio.java
@@ -50,7 +50,7 @@ public class ModeloServicio {
         return precio;
     }
 
-    public void setPrecio(int precio) {
+    public void setPrecio(long precio) {
         this.precio = precio;
     }
